Move config decryption to a new file
..and soup up the tests a bit, to cover MetricConfig and LogsConfig.
